+++ notificare-push/src/main/java/re/notifica/push/internal/NotificarePushImpl.kt
@@ -68,14 +68,14 @@ import re.notifica.push.ktx.INTENT_ACTION_NOTIFICATION_OPENED
 import re.notifica.push.ktx.INTENT_ACTION_NOTIFICATION_RECEIVED
 import re.notifica.push.ktx.INTENT_ACTION_QUICK_RESPONSE
 import re.notifica.push.ktx.INTENT_ACTION_REMOTE_MESSAGE_OPENED
-import re.notifica.push.ktx.INTENT_ACTION_SUBSCRIPTION_ID_CHANGED
+import re.notifica.push.ktx.INTENT_ACTION_SUBSCRIPTION_CHANGED
 import re.notifica.push.ktx.INTENT_ACTION_SYSTEM_NOTIFICATION_RECEIVED
 import re.notifica.push.ktx.INTENT_ACTION_TOKEN_CHANGED
 import re.notifica.push.ktx.INTENT_ACTION_UNKNOWN_NOTIFICATION_RECEIVED
 import re.notifica.push.ktx.INTENT_EXTRA_DELIVERY_MECHANISM
 import re.notifica.push.ktx.INTENT_EXTRA_LIVE_ACTIVITY_UPDATE
 import re.notifica.push.ktx.INTENT_EXTRA_REMOTE_MESSAGE
-import re.notifica.push.ktx.INTENT_EXTRA_SUBSCRIPTION_ID
+import re.notifica.push.ktx.INTENT_EXTRA_SUBSCRIPTION
 import re.notifica.push.ktx.INTENT_EXTRA_TEXT_RESPONSE
 import re.notifica.push.ktx.INTENT_EXTRA_TOKEN
 import re.notifica.push.ktx.logNotificationInfluenced
@@ -84,6 +84,7 @@ import re.notifica.push.ktx.logPushRegistration
 import re.notifica.push.models.NotificareLiveActivityUpdate
 import re.notifica.push.models.NotificareNotificationDeliveryMechanism
 import re.notifica.push.models.NotificareNotificationRemoteMessage
+import re.notifica.push.models.NotificarePushSubscription
 import re.notifica.push.models.NotificareRemoteMessage
 import re.notifica.push.models.NotificareSystemNotification
 import re.notifica.push.models.NotificareSystemRemoteMessage
@@ -101,7 +102,7 @@ internal object NotificarePushImpl : NotificareModule(), NotificarePush, Notific
     internal const val DEFAULT_NOTIFICATION_CHANNEL_ID: String = "notificare_channel_default"
 
     private val notificationSequence = AtomicInteger()
-    private val _observableSubscriptionId = MutableLiveData<String?>()
+    private val _observableSubscription = MutableLiveData<NotificarePushSubscription?>()
     private val _observableAllowedUI = MutableLiveData<Boolean>()
 
     internal lateinit var sharedPreferences: NotificareSharedPreferences
@@ -172,7 +173,7 @@ internal object NotificarePushImpl : NotificareModule(), NotificarePush, Notific
         sharedPreferences.firstRegistration = true
 
         transport = null
-        subscriptionId = null
+        subscription = null
         allowedUI = false
     }
 
@@ -210,10 +211,10 @@ internal object NotificarePushImpl : NotificareModule(), NotificarePush, Notific
             NotificareLogger.warning("Calling this method requires Notificare to have been configured.")
         }
 
-    override var subscriptionId: String?
+    override var subscription: NotificarePushSubscription?
         get() {
             if (::sharedPreferences.isInitialized) {
-                return sharedPreferences.subscriptionId
+                return sharedPreferences.subscription
             }
 
             NotificareLogger.warning("Calling this method requires Notificare to have been configured.")
@@ -221,16 +222,16 @@ internal object NotificarePushImpl : NotificareModule(), NotificarePush, Notific
         }
         set(value) {
             if (::sharedPreferences.isInitialized) {
-                sharedPreferences.subscriptionId = value
-                _observableSubscriptionId.postValue(value)
+                sharedPreferences.subscription = value
+                _observableSubscription.postValue(value)
                 return
             }
 
             NotificareLogger.warning("Calling this method requires Notificare to have been configured.")
         }
 
-    override val observableSubscriptionId: LiveData<String?>
-        get() = _observableSubscriptionId
+    override val observableSubscription: LiveData<NotificarePushSubscription?>
+        get() = _observableSubscription
 
     override var allowedUI: Boolean
         get() {
@@ -316,12 +317,12 @@ internal object NotificarePushImpl : NotificareModule(), NotificarePush, Notific
         val device = Notificare.device().currentDevice
             ?: throw NotificareDeviceUnavailableException()
 
-        val subscriptionId = subscriptionId
+        val token = subscription?.token
             ?: throw NotificareSubscriptionUnavailable()
 
         val payload = CreateLiveActivityPayload(
             activity = activityId,
-            token = subscriptionId,
+            token = token,
             deviceID = device.id,
             topics = topics,
         )
@@ -905,9 +906,9 @@ internal object NotificarePushImpl : NotificareModule(), NotificarePush, Notific
         val device = checkNotNull(Notificare.device().currentDevice)
 
         val previousTransport = this@NotificarePushImpl.transport
-        val previousSubscriptionId = this@NotificarePushImpl.subscriptionId
+        val previousSubscription = this@NotificarePushImpl.subscription
 
-        if (previousTransport == transport && previousSubscriptionId == token) {
+        if (previousTransport == transport && previousSubscription?.token == token) {
             NotificareLogger.debug("Push subscription unmodified. Updating notification settings instead.")
             updateDeviceNotificationSettings()
             return@withContext
@@ -926,17 +927,19 @@ internal object NotificarePushImpl : NotificareModule(), NotificarePush, Notific
             .put("/push/${device.id}", payload)
             .response()
 
+        val subscription = token?.let { NotificarePushSubscription(token = it) }
+
         this@NotificarePushImpl.transport = transport
-        this@NotificarePushImpl.subscriptionId = token
+        this@NotificarePushImpl.subscription = subscription
         this@NotificarePushImpl.allowedUI = allowedUI
 
         Notificare.requireContext().sendBroadcast(
             Intent(Notificare.requireContext(), intentReceiver)
-                .setAction(Notificare.INTENT_ACTION_SUBSCRIPTION_ID_CHANGED)
-                .putExtra(Notificare.INTENT_EXTRA_SUBSCRIPTION_ID, token)
+                .setAction(Notificare.INTENT_ACTION_SUBSCRIPTION_CHANGED)
+                .putExtra(Notificare.INTENT_EXTRA_SUBSCRIPTION, subscription)
         )
 
-        if (token != null && previousSubscriptionId != token) {
+        if (token != null && previousSubscription?.token != token) {
             Notificare.requireContext().sendBroadcast(
                 Intent(Notificare.requireContext(), intentReceiver)
                     .setAction(Notificare.INTENT_ACTION_TOKEN_CHANGED)